requirements.txt
discord
psycopg2; sys_platform != "darwin"
psycopg2-binary; sys_platform == "darwin"
loguru
python-dotenv
dj_database_url
pandas
tabulate
